Add base webpack config
Change-type: minor
