+++ .circleci/config.yml
@@ -376,8 +376,11 @@ jobs:
           at: .
 
       - run:
-          name: Webpack Bundle Report
+          name: Webpack Bundle Report For UI
           command: make build-ui COVERAGE=0 NODE_ENV=test && mkdir -p /tmp/test-results && cp webpack-bundle-report.html /tmp/test-results/webpack-bundle-report.html
+      - run:
+          name: Webpack Bundle Report For Chat Widget
+          command: make build-chat-widget COVERAGE=0 NODE_ENV=test && mkdir -p /tmp/test-results && cp webpack-bundle-report.chat-widget.html /tmp/test-results/webpack-bundle-report.chat-widget.html
       - run:
           name: Generate Coverage Results
           command: mkdir -p /tmp/test-results && tar czf /tmp/test-results/nyc_output.tar.gz .nyc_output && make coverage && cp -rf coverage /tmp/test-results
