+++ webpack.config.chat-widget.js
@@ -7,8 +7,9 @@
 const HtmlWebpackPlugin = require('html-webpack-plugin')
 const path = require('path')
 const DefinePlugin = require('webpack/lib/DefinePlugin')
-const IgnorePlugin = require('webpack/lib/IgnorePlugin')
 const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin
+const mergeConfig = require('webpack-merge')
+const baseConfig = require('./webpack.config.base.js')
 
 const root = path.resolve(__dirname, '.')
 const resourcesRoot = path.join(root, 'lib', 'chat-widget')
@@ -18,74 +19,23 @@ const UI_DIRECTORY = process.env.UI_DIRECTORY || 'lib/chat-widget'
 
 const uiRoot = path.join(root, UI_DIRECTORY)
 const indexFilePath = path.join(resourcesRoot, 'index.html')
-const outDir = path.join(root, 'dist')
+const outDir = path.join(root, 'dist/chat-widget')
 const packageJSON = require('./package.json')
 
 console.log(`Generating bundle from ${uiRoot}`)
 
-const config = {
-	mode: 'development',
-	target: 'web',
+const config = mergeConfig(baseConfig, {
 	entry: path.join(uiRoot, 'index.jsx'),
+
 	output: {
 		filename: 'bundle.[hash].js',
 		path: outDir,
 		publicPath: '/'
 	},
 
-	resolve: {
-		extensions: [ '.js', '.jsx', '.json' ],
-		alias: {
-			'@jellyfish-ui-components': path.resolve(uiRoot, 'components'),
-			'@jellyfish-ui-shame': path.resolve(uiRoot, 'shame')
-		}
-	},
-
-	module: {
-		rules: [
-			{
-				test: /\.(js|jsx)$/,
-				exclude: /node_modules/,
-				use: [
-					{
-						loader: 'babel-loader',
-						options: {
-							presets: [ '@babel/preset-react' ]
-						}
-					}
-				]
-			},
-			{
-				test: /\.css$/,
-				use: [ 'style-loader', 'css-loader' ]
-			},
-			{
-				test: /\.(woff(2)?|ttf|eot|svg)(\?v=\d+\.\d+\.\d+)?$/,
-				use: [
-					{
-						loader: 'file-loader',
-						options: {
-							name: '[name].[ext]'
-						}
-					}
-				]
-			}
-		]
-	},
-
-	devtool: 'source-map',
-
 	devServer: {
 		contentBase: outDir,
-		compress: true,
-		port: 9100,
-		historyApiFallback: {
-			disableDotRule: true
-		}
-	},
-
-	node: {
-		fs: 'empty'
+		port: 9100
 	},
 
 	plugins: [
@@ -106,25 +56,16 @@ const config = {
 				VERSION: JSON.stringify(`v${packageJSON.version}`)
 			}
 			/* eslint-enable no-process-env */
-		}),
-
-		// The moment.js package includes its locales by default, they are huge and
-		// we don't use them, we're going to ignore them
-		new IgnorePlugin(/^\.\/locale$/, /moment$/)
+		})
 	]
-}
+})
 
 // eslint-disable-next-line no-process-env
-if (process.env.NODE_ENV === 'production') {
-	config.mode = 'production'
-	config.optimization = {
-		minimize: true
-	}
-} else {
+if (process.env.NODE_ENV !== 'production') {
 	config.plugins.push(
 		new BundleAnalyzerPlugin({
 			analyzerMode: 'static',
-			reportFilename: '../webpack-bundle-report.chat-widget.html',
+			reportFilename: '../../webpack-bundle-report.chat-widget.html',
 			openAnalyzer: false
 		})
 	)
